backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
torch==2.3.1
transformers==4.42.3
requests==2.32.3
